app/requirements.txt
python-dotenv
fastapi[all]
pymongo
pg8000
sqlalchemy==2.0.35
passlib
bcrypt==4.0.1
python-multipart
python-jose[cryptography]
pypdf
langchain
langchain-mongodb
langchain-community
langchain-google-genai
cloud-sql-python-connector
google-auth
faiss-cpu
pandas